Mail: Fix "Save address" as new Person file
If a Person file with the selected email address doesn't exist yet,
launch the People app with a BMessage "M_LAUNCH_PEOPLE" that triggers
the People app to create a new Person file.
The message includes "META:email" and "META:name" for email address
and contact name. The contact name may be empty, if it wasn't included
in the mail's from/to/cc etc. field.

Don't show error alert, if People was already running.

Fixes #14766

Change-Id: Id4f7ae50aa1ea283974f2fa067ac9e199f369db5
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7084
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>