ffmpeg: Increase accuracy of frame count computation.
BePac Deluxe checks the reported frame count against the actually
read frames count, and if the two do not match, it considers
this an error and bails out. So, we need to be as accurate as
possible here.

Additionally, fix Duration() for when start_time is nonzero,
and stay in integers where possible.

Fixes BePac Deluxe (again.)

Change-Id: Ic063fc15215dbfe3437d2430358227d7e2fe17fc
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7422
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>