Tracker: Offer "Get info" when a symlink fails
While it's obvious which link is broken when the user double-clicks it,
it isn't when the link was tried to be opened by a script, or the
launch_daemon etc.

In case resolving a symlink fails, offer a button to call Tracker's
"Get info" panel with the broken link that has info about the link and
its target.

Change-Id: Id878d0ea3da0db65ade21dd4656c18ecba585784
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7434
Reviewed-by: waddlesplash <waddlesplash@gmail.com>