TCP: Let EndpointManager set Local and Peer addresses.
The Local address will apparently be set by the module->bind() call
in EndpointManager::_Bind(), while the Peer address will be set by
EndpointManager::SetConnection().

This means that if BindChild() fails for some reason, we will now not
be left in a state where the LocalAddress is set but we are unbound.
That could be the cause of #17058, as that panic occurs when
TCPEndpoint::IsBound() returns true, but we have not been added to
the bound hash.
Change-Id: I3121d0d99289d8cf83c7ed3c3e2d413feb6dad70
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7490
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: Jérôme Duval <jerome.duval@gmail.com>