ffmpeg: implement equivalent of avcodec_guess_frame_rate
The frame rate in the codec context isn't always a frame frate in the
video sense. It may be a smaller time unit used as a base for the frame
presentation timestamps.

I did not initially notice this because we set it to a default value
from the AVFormat passed input encoded_video format, but it is erased
after reading the first frame.

Fixes #18797.

Change-Id: I6157ae9942f6d5ca32aad7bac012df0822659a21
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7563
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>