BFilePanel: fix items from favorites menu not working
Regression from hrev57294.

The code introduced there relies on the state of the "open" button being
updated from a previous selection in the file panel. But that is no the
case if the B_REFS_RECEIVED message does not come from the file panel
(in this case it comes directly from the favorites menu).

Fixes #18627.

Change-Id: Ic12b1989a4fbd631ce048aa6286d4d7c60ab1d99
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7577
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>