ioapic: more useful logging
Back when this log was introduced, it could not be tested because no
hardware hitting that limit was known. Now there is some.

Add the actual number of IOAPIC redirection entries to the log, so we
can know how many redirection entries we need, and also add a missing
newline.

Change-Id: I00fa0c21628fc336c364cd975dd48801add75302
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7579
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>