usb_hid: add quirk to enable 6th button on Elecom M-XT4DR trackball
Linux implementation: https://www.spinics.net/lists/kernel/msg2665216.html

In our case it is not possible to read the original descriptor from
quirk handlers, so, I had to hardcode it all. The descriptor is the same
as the original one with just 3 bytes changed (detailed in comments).

Other Elecom hardware requires similar quirks, but can have different
number of buttons, so I did not add it for now.

Change-Id: Ie24c8535f3b4535eb414a4e35a62696ea1f9056d
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7580
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>