Fix printf format warnings in B_TRANSLATE strings
Previously these were not checked by the compiler due to a missing
function attribute.

Unfortunately, for the translation macros to work, the translated string
must be identical in all architectures (otherwise, we would need
different translations and different catalogs for each). This means the
B_PRI* macros can't be used, and instead the parameters must be in one
of the types handled directly by printf. Change the variables types
directly where it was easily possible, and otherwise, use a cast.

Change-Id: Ib77a7e378b7c508f6e7a015bbe3cbb4c2c096bfa
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7665
Reviewed-by: waddlesplash <waddlesplash@gmail.com>