syslog_daemon: add syslog_max_history option
This driver setting in ~/config/settings/kernel/drivers/kernel allows
setting an arbitrary number of syslog files to keep around.  The
files are now named with a numeric suffix instead of being called
"syslog.old".  The old files are rotated so that syslog.1 is the
newest, then syslog.2, and so on.

Change-Id: I7e3d01caa34680553b161ceb27ab8c3b2eaf508c
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9055
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>