kernel/fs: Fix readv/writev after the seek-disabled refactor.
This fixes a regression introduced in 34fcf3d9ead1371b57e46877071ced38235e789e.

That commit correctly adjusted the "pos" checks at the top of these
functions, but missed that there was a place in the loop where pos
is incremented. After the first increment, we would have a non-zero
pos, and so the new checks in the actual read/write routines would
return an error, meaning that only the first iovec was ever processed.

This fixes WINE following the aforementioned change.