kernel/events: Fix wait_for_objects.
In the preceding commit, I missed that "events" was cleared
inside the select loop, making a use later on useless.

Now we don't clear it in the loop, but wait till the end
to do so. (Considering we return without clearing it in
case of out-of-memory or other problems, this may not
even be necessary, but it at least preserves existing
behavior.)