kernel/user_debugger: Clean up scheduling of flushes a bit.
Instead of doing the checks in profiling_do_sample, we now do them
in profiling_event. This allows some comments to be consolidated
and make the logic clearer (but it should be functionally the same.)

Also add a check that profiling is still happening at the end
of profiling_flush.