libsolv: Import 0.3.0_haiku_2014_12_22.
As discussed in #18339, it seems to make sense to vendor libsolv,
given how closely tied it is to the package kit, the difficulties
of upgrading it without breaking anything (even when using SOVERSION
to avoid conflicts), and other concerns.

We already vendor libraries much larger than this (e.g. Zydis),
so vendoring libsolv seems fine.