TCP: Do not start the TIME_WAIT timer in Free() if we are about to delete.
The TIME_WAIT timer, when it fires, either sets FLAG_DELETE_ON_CLOSE
(if FLAG_CLOSED is not set), or releases the socket reference directly
(if it is set.) As we set FLAG_CLOSED in Free(), we only want to start
the TIME_WAIT timer if FLAG_DELETE_ON_CLOSE is not set yet.

Probably harmless since the timer was cancelled on deletion anyway,
but it doesn't hurt to be clearer here.