BTimeSource: Protect against values too large to fit in a float.
This can only happen if the real time or performance time values
specified are very large (more than 24 bits), which should only
happen if the time specified is system_time() and the "last" time
is 0. Under that circumstance, last_drift should be 1.0f,
so we can avoid using it at all. Otherwise, invoke debugger().

This would have caught some (but not all) of the problems fixed
in preceding commits.