New Incremental Search Panel Inside The results Window
This commit has all the work that I have done on the incremental search panel. I am yet to make the code in a better way by using better practices but I intend to do that over the next few days. At the moment, my buildtools on a clean repo is still building. As soon as this is over, I will refactor the important parts of this in an attempt to do better memory management, etc.

I have also tested this with the lib_debug.so as stated in the email and fixed the issue that was causing the frequent crashes.